Finished the DC corruption fix.
I could not find the source of the slowness in the hangar. But quitting
the mission and restaring it seems to fix things. I wonder if it's some
sort of cache thing? The DC corruption bug has been fixed; but I still
think the names don't need to be copied all over (a single copy should
suffice, or maybe the names can be released once the textures have been
tagged). I'll leave that for later.