Fixed the bloom being blocked by the HUD and lasers.
The problem was that the bloom mask must use the right alpha. The
default mask should be float4(0,0,0,0) and then use the alpha to
blend with the existing mask. Then, when doing the bloom itself,
ignore the alpha completely and set it to 1. This seems to fix
most of the observed artifacts.