Split the PixelShader code and constant buffers. Fix performance.
This shader was too big. Made two new pixel shaders: one to render
the Dynamic Cockpit, and one to render the HUD foreground/background.
The PixelShaderDC can probably be split a bit more. This should improve
performance a little bit since the simplest PixelShaderTexture will be
used most of the time.

The constant buffers have also been split. The DC-constants are now on
slot 1 and are only written when DC-related stuff is rendered. The other
constants remain in slot 0.

Splitting the PixelShaderTexture seems to have fixed the performance
problems when the hook_time was disabled and DC was enabled: I think the
pixel shader was just too complex and maybe all of the code paths were
being executed even on non-DC textures.