Hyperspace tunnel experiment, tone map compression.
Added code to enable the new hyperspace tunnel in the future. The mouse
can now be used to build a Full Transformation Matrix (looks like the
center of rotation may not be at the origin).

Added basic tone map compression in the bloom shader. The formula is
now based on the HSV space (V is multiplied by bloomStrength and then
tone map compression is applied before converting back into RGB).

Started moving the code towards the pyramidal blur.