Added bloom passes per level and external camera fix for DC.
Each level in the bloom pyramid can be performed multiple times and
this number is now configurable through an external file.

DC is now disabled in external camera mode -- this prevents some
flickering when HUD elements are displayed in-cockpit and then the
external camera is activated -- this may need a bit more testing
since I'm not sure the current player is in slot 0.