Numerous fixes and additional bloom parameters.
Fixed a small artifact in DC. The shields did not erase the last
line and a small line was visible on the left side.

Added independent params to bloom lens flare and suns.

Added param to reduce the strength of the bloom inside the cockpit.
Updated the PixelShaderTexture and PixelShaderDC shaders accordingly.
Removed the colorMul setting.