Added kernel of size 3 and explicit ZBuffer state management for Bloom.
The bloom shader now disables ZWrite before bloom is rendered -- this might
fix a few reported artifacts.

Added Gauss kernel-3 for bloom; but disabled it. It works! But I think I
prefer the kernel-5. To enable it, just change the horizontal and vertical
gaussian loop to go from 1 to 2 and enable the relevant kernel in the common
bloom file.