Added debug code for Bloom and minor fixes to shaders.
The alpha channel in the bloom shaders was multiplied by bloomStrength
and was also part of the tonemap compression. There is no need to do
that and maybe it was causing bugs. I'm only using the rgb components
now.

Pressing Ctrl+Alt+D will dump a few buffers to help debug the Bloom shader.