Fixed black artifacts in Win7 and enabled Reshade again.
The black artifacts where happening because I removed the "discard"
instruction in MainPixelShader.hlsl. On Win10 this had no effect; but
on Win7 this rendered black pixels where nothing should've been displayed.
This shader is activated in the Tech Room when the overlay is displayed on
top of the 3D floating object; but it is also used in-flight when the
targeted object has a sub-component that can be selected. The sub-component
is rendered with a small green box on it, and this box is rendered with
the MainPixelShader! That's why it didn't match the 3D perspective when
run in VR mode.

The discard instruction has been restored and a new shader (BasicPixelShader)
is now used where the MainPixelShader was being co-opted (in steamVRResize, I
believe). No artifacts appear in Win7 and it can even run along with ReShade.

More work is needed to fix the green sub-component bracked in the targeting
computer for VR.