Several fixes to SSAO (bloom is now applied after SSAO).
Reordered SSAO so that it's executed before Bloom and hence Bloom is
applied on top of SSAO -- may have to double-check that later.

Added more parameters to ssao.cfg to control the zoom of the effect
and disable/enable blur.

Added Ctrl+Alt+O to toggle SSAO and Ctrl+Alt+D to display the SSAO
debug buffer.

Fixed SSAO in the external camera (just resolve the Depth Buffer right
before the SSAO is computed if it wasn't resolved during the regular
Execute()).