Added second depthBuf for background objects and texture flags.
Added code to D3DTexture to distinguish between cockpit textures and
external player views. Added another depth buffer, SRV and RTV so that
rendering the pos3D goes to a different buffer depending on what is
being rendered. The cockpit is rendered at the end; but unfortunately
the external view renders the player's craft first and then the background,
so we end up with a hole in the background depth buffer.

I may have to roll back these buffers; but I can try and see if it helps.
The alternative is to modify the SSAO shader to ignore distant samples
after a certain threshold.