Experimental version of shading through Bent Normals
Enabled BentBufR even when SteamVR is not active (will be used as
temp buffer to blur the bent normals). Bent normals are computed in
a hacky way; but it kind of works.

Also added matrix constant buffer in the pixel shader to transform
the light vector. That part needs to be checked because I'm sure
it's not 100% correct.