Fixed a few minor bugs in SSAO. Removed some unused params.
There was bug in the ssao blur (I had to clear the buffers) and
the normals buffer was not being set properly. The SSAO blur now
doesn't show halos in high resolution and only minor halos in
medium res.

Almost ready to release -- just need to fix the SteamVR path as
well.