Added diffuse buffer and direct illumination SSDO shader.
The diffuse buffer is probably not going to be useful; but the
infrastucture can be used later to hold the direct-lighting
framebuffer for the second SSDO pass.

Finally fixed the SSDO shader for direct lighting. It's still
experimental but it looks much better than regular SSAO. I'm also
using the code that I wrote for bent normals for SSDO -- which
is also experimental. But I think it's worth continuing on this
path.

Also added debug parameter and lighting direction in the SSAO.cfg
file.