Removed all Diffuse Buffers, SRVs, RTVs. Moved ssaoBufR to normal path.
In other words, the ssaoBufR, SRV and RTV are now initialized even when
SteamVR is not active. This is necessary because ssaoBufR will be used
to hold the SSDO Indirect Illumination.

The SSDO Direct illumination can be computed at half-scale in this
version; but SSDO Indirect scaling needs to be checked.