SSDO Indirect is fixed now and it can be blurred.
The problem was that offscreenAsInput is used as temporary buffer to do
the blur; but I had to resolve the offscreen buffer again. Also, the
indirect ssdo buffer (ssaoBufR) has to be cleared on every frame.

This version is working; but I still need to fix the scale (only direct
ssdo can be scaled down right now).