SSAO Transparency fix for Yavin and other stuff.
Changed how lights are aligned so that yaw,pitch,roll is mapped to
the Z+ axis. It works; but needs more more testing to get it working
properly.

Yavin IV trees now map alpha to solidity for SSAO.

Added Moire offset for SSAO, it works; but I think I'm going to need
something more to fix SSAO on the flat surface of Yavin IV.