Added MipLevels to SSDOIndirect, changed weight function.
The weight function is now dependent on the radius, so it fades
to 0 on max_radius.

Also reverted the ComputeRotationMatrix function to the old version;
but now I'm seeing "gimbal-lock" issues with the light. Maybe this is
happening because I also changed how the BentNormal.z is computed in
SSDODirect?