Restored SSAO to the way it looked in 1.0.8, added miplevels and moire offset
Flipping N.z in the pixelshader had a negative effect in the way SSAO was
rendered. I had to stop flipping N.z to restore SSAO to the way it looked
before. However, I added miplevel sampling to SSAO and moire offset (this
offset can actually be set to 0 externally).