Use TBN Matrix to compute light position from yaw, pitch, roll.
I ended up using a TBN matrix to solve this problem. I'm using spherical
coords to align the yaw,pitch,roll with the Y+ axis. This axis becomes N
and then it's trivial to build B and T. In this frame, I'm applying the
roll about the Y axis to the T and B and then I'm computing the inverse
matrix to transform the light into view space.

The light_vector should be set as follows:

light_vector = 0, -1, 0

0, 1, 0 also works but it tends to put the light below the "horizon".