The edges of the effective viewport are now darkened for SSDO.
I'm computing the edges of the effective screen in non-VR mode. If
a sample lands outside these edges, it's considered occluded. This
prevents halos around the edges and it doesn't look that bad when the
surface is illuminated either.

However, the blur may have to be updated to see these edges too.