Added fadeout for SSDO, tweaked bloom mask.
For distant objects, SSDO quits abruptly. I addded a fadeout range
from 10km to 15km. This is most notorious in Yavin IV.

The bloom mask is used to inhibit SSDO; but sometimes it inhibits too
much, I added a threshold to cut the bloom mask at 0.95 and removed
it's effect in the SSDOAdd shader. Maybe I should add it back later?