Added plumbing for future HDR format, bloom threshold set at 0.975
The bloom mask for the engine glow is messing up SSDO. I changed the threshold
to 0.975; but I'm still seeing artifacts.

Also changed the format for ssaoBuf/SRV to RGBA16F to support HDR in the
(near) future. I had to change a lot of the plumbing for these buffers as
well as the bloom mask buffers/SRV/RTVs -- I noticed that these resources
should be created when either bloom or ssao are active.

SteamVR has not been tested after this commit. Consider this commit
suspect if things crash.