Release 1.0.9 preview.
Fixed ssaoBuf at 16-bit when not in HDR mode. If the dot product is not
saturate()'d then it will try to apply a 16-bit HDR buffer on top of
an 8-bit buffer without tone mapping. Interestingly enough, this actually
looks interesting... almost like HDR. But no, this is a bug.

Removed some debug code to enable the pre-release.

Added some default parameters to remove them from the config file.