Further improvements for the shadows.
Switched to flat rendering so that debugging shadows was easier.
The shadows don't look that bad and I think they even have some
soft edges; but there are still large artifacts to be fixed.

Shadows can be toggled with Ctrl+I and only work with in the
Bent_Normals mode.