Ray-traced shadows are working properly now.
Things that stil have to be fixed:

1. Shadows should fade with distance to hide clipped shadows
2. Fix shadows on distant objects.

I removed one condition from the code that was causing a lot of
artifacts (diff > 0 && diff < 0.0) --> return 0; Without this,
the shadows look fine.

Shadows are not perfect, we can see they are coming from clipped
geometry and they disappear abruptly at the edges of the screen.
However, we get nice hard-to-soft shadows with a single ray!

The shadows can be rendered to their own buffer to blur them, or
maybe we can add them to the ssao buffer.

The specular highlights didn't work very well and I don't know why
yet.