HyperCompose shader added. Debug commit: the hyperspace effect is broken.
In this commit, I've added the "hyperzoom" effect and a HyperCompose shader.
When exiting hyperspace, I'm going to blend the background (everything except
the cockpit) with the exit streaks. The background will be zoomed-in and blended
with the streaks while the cockpit is rendered on top of everything. To achieve
this, the cockpit draw() calls are redirected to shaderToyBuf when exiting
hyperspace.

This is a debug commit, the effect is quite broken at this point and the cockpit
doesn't have transparency. I still have to fix the transparency on the cockpit and
blend with the hyper-exit shader.