Fixed the transparency problem in the cockpit when doing hyperzoom.
The problem was that light textures in the cockpit overwrite the alpha
already present in the _offscreenBuffer. The proper fix is to change the
alpha blending state; but if I do that, I'll probably break something else.

Instead, the workaround was to disable all cockpit light textures when exiting
hyperspace. We'll lose some detail when exiting hyperspace; but it's only a
few frames, so this is a much safer option.

Everything is ready to blend the effects now.