HyperExit is now blended with HyperZoomCompose; still needs debugging.
The HyperZoom is now blended with the HyperExit; but it won't work on an
actual hyperjump. It actually seems to revert a bit when the blend should
happen. #define HYPER_OVERRIDE on Direct3DDevice.cpp and PrimarySurface.cpp
to see the blending in action (some artifacts may happen because the effect
is rendered outside an actual hyperjump, though -- the cockpit will
disappear).