Fixed the black gap when exiting hyperspace; but more bugs need fixing.
The gap was happening because there's a point where no trails are rendered
and I was using those trails as an indicator of when to render the new
hyperspace effect. The fix is to keep rendering the new effect even when
no trails are visible. This gap only happens during HS_HYPER_EXIT. However,
the problem now is that no cockpit is rendered because of more technical
reasons.

I need to re-do the logic that hijacks the hyperspace effect draw calls and
probably capture the cockpit at *all* times during the hyperspace jump.