Hyperspace Effect Fixed. The sequence is fully-blended into hyperzoom.
There might be a bug, though: I'm re-using the shadertoyAuxBuf on the
second pass (the pass where I'm rendering the cockpit). This buffer has
the background that is captured during entry and exit, so by overwriting it
I'm probably destroying its contents during entry. The exit doesn't suffer
from this problem because this buffer is captured anew for every frame after
exiting hyperspace.

I also need to fix the lighting and make it rumble; but... details.