Hack to enable true translation inside the cockpit (better 6dof).
Instead of adding the translation to the view matrices, let's alter
the cockpitReference values in PlayerDataTables so that polygons don't
get clipped anymore. This way, we can't "lean outside" the cockpit; but
the world won't respond to 6dof. Probably a nice tradeoff?

Only the DirectSBS mode is affected in this commit.

This commit also breaks the shake during hyperspace. It will also break
the shake from engine wash.

The translation is also a bit shaky; but it doesn't shake if done inside
the CockpitLook Hook.