Fix for DTM's skydomes, new bloom setting, cockpit lean code removed.
Added a new setting to change the bloom of the skydomes in DTM's
maps. Fixed the ugly shadow that appeared in the distance (it was
an "if (p.z > INFINITY)" test in SSDOAdd). Added new flag to mask
the skydomes. Refactored all INFINITY_Z constants across all shaders
and replaced them with INFINITY_Z1.

The Yavin map now looks OK after this change.