Fixed the back/projection code so that intersections now make sense.
I was using 0.5 to center the 3D coords; but it should've been 1.0 instead.
This happens because the vpScale is multiplied by 2 in the nonVR path.
For VR, the factor is 1, so I'll need to use 0.5 again when doing projections.
The shader was also modified to display 2D coords directly to avoid additional
complexity.

When using the external camera, the screen goes black when I press the ESC
key -- need to fix that bug.