Ray-triangle intersection is now working properly.
I had to hack together a new ray-triangle intersection function that
combines the Moller-Trumbore algorithm to find t/P and then I used
the old function to compute the right barycentric coords. For some
reason the Moller-Trumbore algorithm did not provide sensible UV
coords, so I was forced to do this.

On the brighter side, the intersection seems to be working well now.