Upgraded OpenVR version. Plumbing to support AC in DirectSBS mode.
Nothing much changed in OpenVR; but I'm still using the newest version.

Moved the code around for RenderLaserPointer since it didn't work well
for DirectSBS. Also changed the vertex buffer for the barrelVertexBuffer
because the other one was getting transformed and that changed the size
and aspect ratio of offscreenBuffer.

project/backproject doesn't work in DirectSBS mode. Need more testing.