DirectSBS project/backproject works properly. AC is working.
DirectX divides the PixelShader output by w internally. We don't see
that division; but it happens, and it must be done explicitly in C++.
After that division, the coordinates are in DirectX format (-1..1) so
that has to be transformed to post-proc (UV) coords (0..1) and the
projection finally matched what we see in the screen.

AC is now working properly in DirectSBS and non-VR modes.