Hyperspace Pause and HMD-compensated Laser Pointer.
The laser pointer's position and rotation is now compensated against
HMD rotation and translation. In a few words, the laser pointer should
stay in the same position if the HMD moves. I propbably should make
this a toggleable setting for people that would like to "point-and-click"
using gaze.

Also added experimental "Hyperspace Pause" feature.