Focal Distance is now used in reconstruction and AC. New Params.
Looks like the focal distance was the factor that was missing in the
project/backproject code. With this new factor, the yaw/pitch HMD motion
is compensated much better. It's not perfect because the pitch axis does
not cross the center of the screen; but it's really good and now things
make more sense.

The stereoscopy is probably messed up now and the METRIC_SCALE_FACTOR will
probably have to be divided by g_fFocalDist to keep the current stereoscopy.
Otherwise, things will get more stereoscopic depth when we increase the focal
distance.

Added debug AC param to display the triangle where the intersection
was computed. Added "compensate_HMD_motion" to toggle the HMD-compensation
so that people can use this either in VR mode or to point with the gaze.