Changed AC actions from strings to WORD scan codes.
This way, no translation/interpretation has to be done at run-time. The
translation is done when the AC commands are loaded and it's only done
once. The strings are translated to scan code sequences that are just
executed at run-time.

Only basic support for alphabetic keys is working right now. More work is
needed here.