Fixed DC coords above 1 by using frac()
If the UVs for a DC element are above 1, the UVs won't be wrapped and
the DC element will be blank. I fixed this by using frac(input.tex) so
that it wraps around the 0..1 range.

Also commented out the code that uses the hook_normals hook so that this
ddraw can be released without using that hook. I'll re-enable it later.