Added full_cockpit_test, laser_pointer_length & debug_laser_dir.
full_cockpit_test allows the user to force the intersection test on all cockpit
surfaces. This makes it easier to point with the laser pointer.

laser_pointer_length lets the user specify the distance from the controller's
origin used to compute a fake intersection.

debug_laser_dir is an internal setting that changes how the laser pointer direction
is computed.