Added support for Gamepad buttons and axes coming from PSMoveService.
Fixed bug in the SPACEBAR keycode. It was still sending an "S" because
I forgot to `return` from that case.

Also split the HMD motion compensation into rotation and position.

I still need to map the left/right buttons in the Joycon to something
useful like, maybe "1/3 speed" and "match speed" perhaps?