Fixed AC for the new Y-Wing cockpit. Arrow keys are also working now.
For the Y-Wing, I needed to invert the y-axis in the AC definition file;
and add 1.0f to the negative coordinates until they became positive. In
other words: the AC file must have inverted y-coords; but negative UVs
must *not* be mirrored. This works both for the Y-wing and A-wing AC
cockpits.

To fix the arrow keys, I needed two things:

1. Add the KEYEVENTF_EXTENDEDKEY flag to the scan codes.
2. Send all the scan codes in one SendInput instruction, including the
   key down and key up events.