Release 1.1.0. Various fixes for the Hyperspace effect.
The illuminated textures are now rendered properly during the hyperspace
effect (I was just skipping them previously). The fix was to discard the
parts that should not bloom in PixelShaderTexture and to set alpha = 1.0
in PixelShaderDC.

Also fixed the bloom params for the hyperspace effect. They were being
ignored.

Tested the hyperspace effect on Win10 and Win7.