Material lists can now be loaded. HUD fixes for glass specularity.
Material lists can now be specified in lists of the form:

[TEX000XXX,TEX000YYY,...]

There's a limit of 256 chars, and I may need to revise how the names (char*)
are copied; but it looks like it works.

I also increased the bloom for glass surfaces. This seemed to have a negative
effect on the HUD as well, so I'm increasing the "shadelessness" of the HUD
to compensate. I've also observed some weird bloom artifacts if I set
spec_int and glossiness to 0 wheb bIsAimingHUD is true. Not sure what this
means; but this commit looks fine.