Custom FOV and Metric Multiplier can be applied and saved.
The custom FOV can now be set in real-time using Ctrl+Alt+Shift-Right/Left
arrows. The settings are automatically saved to FOVParams.cfg and loaded
after the first frame is rendered on subsequent runs. If the file is not
present, then the default game FOV is used.

The MetricMult param is used in the SBSVertexShader vertex shader to increase
and decrease the scale of the reconstruction. The net effect is that stereoscopy
is boosted or decreased. This setting is also saved automatically.

Pretry sure the MetricMult will break the Active Cockpit laser pointer; but
it should be easy to fix.