DC should not be reset when textures are freed, only when the cockpit changes.
Depending on which hook is installed, the textures are either freed-and-reloaded
or just freed. By resetting DC when textures were freed, I was breaking DC for
people with the hook_60FPS installed. This problem doesn't happen with hook_time,
though.

The proper fix is to keep track of the cockpit name. If it changes, then we
reset DC and reload it.