Minor improvements to SSDO Indirect, improved limits tests using any()
I'm now using any() and all() statements to perform UV limits checks in
a number of shaders. This should improve performance a little bit.

Minor improvements to the SSDO Indirect shader. For starters, I've
re-enabled it; but there are large areas that get indirect lighting
where none should be. Still trying to figure out why -- probably
because of self-occlusion.