SSDO Indirect is looking much better now:
TODO: Fade the effect at the end of the radius (it cuts off rather abruptly
right now.

TODO: Gray light reflected off a gray surface should not paint objects in
gray color. Instead, the gray color should be mixed with the albedo and then
that should be added on the direct light component.