Emissive Materials test.
It's possible to add an emission mask to store additional lighting
contribution from emissive materials. However, it's noisy (maybe
because emissive materials tend to be small in size), we don't have
the 3D position (because we don't want lasers to cause occlusion)
and it's only a subtle effect anyway (because it's computed during
SSDO, it only affects the vicinity of the emissive materials/lasers
so it's only visible if I turn bloom off). Probably not worth it.

It might be better to store the N closest lasers to the cockpit in
view coordinates and use them as light sources. However, I'm leaving
this commit anyway in case I want to revist this in the future.