Semi-Fix for Active Cockpit.
I broke AC a while ago when I changed how the limits where checked.
I had to use float2 instead of float. This fixes the shader; but
AC doesn't detect all the intersections and thus misses several
buttons. More work is needed in this area; but the good thing is
that the intersections that are detected work properly.