Fixed LaserPointerVR shader, reverted default FOV to 1.0
If the FOV is different from 1.0, that'll break AC as many intersections
won't be caught. AC still works in release 1.1.0; but became broken in
1.1.1. If I set the FOV to 1.0, that fixes most of the buttons except
for the ones on the left text box in the A-Wing. So it's not a full fix;
but it's a strong hint.

Also, the limits for the LaserPointerVR were float's not float2's, so
this shader just presented a black image.

I really need to go back to the FOV problem later and make it work with
the game's FOV.