Fixed AC for hook_d3d, non-VR case.
I did not have precision issues, I was erroneously multiplying by
metric_mult in the non-VR code path. Removing that fixed all remaining
issues. The code works even if the user changes the game's FOV. I even
re-introduced g_fFocalDist again into the calculations and it works.

Of course, g_fFocalDist should be related to the in-game's Focal Length;
but that's a different issue.