Fixed light textures, background detection, brackets in VR + hook_d3d.
Still need to test that the proper stereoscopy is applied; but the
detection seems to work.

Fixed empty DC elements while in hyperspace (they showed the effect
through the cockpit).

Also added std::vector of textures to link color and light textures.
I disabled it; but the linking seems to work fine. I may come back
to this in the future to remove unnecessary draw calls.