Fixed Deferred shader; added laser lights.
I broke the deferred when I added the bent normals texture for SSDO
(the bent normals are needed to compute the contact shadows). This commit
fixes the Deferred shader.

I'm also adding a new feature: the closest lasers to the camera are stored
in a list that then gets sent to the shaders to use as point lights. They
still need to be refined; but it seems to work.