Lasers look better but need more work...
The light should be attenuated with depth, so that they don't abruptly
disappear.

Need general intensity setting for the lasers so that we can have a large
distance effect without making a huge glow.

The glow seems to move along with the craft orientation, instead of "sticking"
to the laser position -- no clue why this is happening.