Removing harsh transitions around edges of ships.
Even with MSAA on, there are some hard edges around the ships where
they transition into the background. The background is shadeless, so
this transition was hard. Here I'm trying to make it softer. One transition
that I could not remove was this test in SSDOAdd.hlsl:

 if (Normal.w < 0.01) {

Removing that messes up the bloom mask and I don't know why. I've made
this transition softer in several places, though.

This same work has to be repeated for the SSAOAdd shader