Release 1.1.2.
Minor fixes for laser light bloom, cockpit inertia beta support.

The bloom effect seems to interact somehow with the dynamic laser lights,
I moved the place in the shaders where the bloom is written to try and
avoid this interaction.

Ctrl+D now toggles the laser lights.

We need to clear the captured image when entering hyperspace if cockpit
inertia is enabled. Unfortunately, this is a toggleable variable in
the hook, so I need a better way to solve this problem.