Fix to use 32-bit indices when D3D hook is present.
DTM reported a very insidious bug: if the D3D hook is enabled, then we may not
use the right indices to capture the HUD elements for DC when we load missions
with lots of geometry. The bug can be reproduced with the "Enemy at the Gates"
mission; but we have to be using the external camera when taking off from the
hangar.

For missions with low (<65K) geometry, this was invisible because 16-bit indices
were enough. But in missions with high poly-count the 16-bit indices would wrap
around and Hilarity Would Ensue.

The fix is to judiciously switch to using 32-bit integers when the D3D hook is
present and use the index buffer directly. There are at least 2 other functions
that have to be fixed in the same way.

I also added a new shortcut: Ctrl+R to reset the HUD limits so that they get
re-captured. This seems to help fix artifacts sometimes.