First steps towards adding XWA Lights to the New Shading System.
In Skimish Mode, looks like the light at index 1 is the main light as it matches
the position of the Sun/Nebula and it seems to have the highest intensity.
However, in regular missions, the main light seems to be at index 0. Looks like
we'll need to find the light with the highest intensity and put that in a
different slot to use it for SSDO.

This commit uses the HeadingMatrix (the same used for the Hyperspace effect and
5dof code in CockpitLook) to transform the light positions into viewspace.
However I still need to compensate for cockpit camera yaw/pitch. Otherwise,
looks like this is working.