XWA Lights and hook_exterior fixes.
Restored 8 XWA lights. Added xwa_lights_saturation and
xwa_lights_intensity.

Fixes for the hyperspace effect when the hook_exterior is
present.

Fixes for the HUD material when the hook_exterior is present
(it is no longer affected by SSAO/SSDO shadows).

The aiming reticle looks better in the exterior HUD shader;
but Jeremy's external hook just made that obsolete.