Fixed white reticle when the HUD is hidden.
if g_bDisableDiffuse is set, then the reticle will go white when the HUD is
fully hidden and there are no brackets on screen. Brackets are used to tell when
the HUD/GUI has started rendering; but if they are not on-screen, ddraw will
mistakenly think the reticle is part of the 3D contents and removes the shading
from it.

There are two ways to fix this. In this commit, I changed how g_bStartedGUI is
computed by adding bIsAimingHUD. This might break VR, though. If that happens,
then we need to add !bIsAimingHUD to the section where g_bDisableDiffuse is set
to 1 to prevent setting it for the reticle.