Switched to use the sun texture's center in pixels.
For some reason, using the light direction to add flare didn't work well. Many
missions showed flares on top of planets or flares going in the wrong direction.
Here I switched to compute the centroid of the sun texture and that always
places the flare in the right position; but the centroid can only be computed
reliably when the sun is fully visible... Still needs a lot of work.