Using barycentric coords to compute the center of the sun texture.
This version works fairly well; but the center still jumps a little bit in
certain places and it's not 100% centered all the time either. The good news
is that the light vector is now very close to the center, so it should be
possible to associate lights.

TODO: Do I need to this to be perspective-correct, or should I back-project and
find the barycentric coords in 3D? Also, the sun texture is at ~40m away; but
we push it back in the DirectSBS version (maybe in the non-VR path too), that
may explain the difference between the light's position and the Sun's center.