Interpolating the sun's centroid in metric 3D to make it perspective-correct.
Instead of using the barycentric (u,v) to compute the centroid of the Sun in
2D coordinates, let's back-project each vertex of the containing triangle into
metric 3D first and then do the barycentric interpolation in 3D. Then, the
resulting centroid can be projected back into 2D to make it perspective-correct.
This version seems to put the centroid at the right spot all the time.