Refactored code, made space for sun color in PixelShaderCB
Refactored the PixelShaderTexture constant buffer for all the shaders that share
it (PixelShaderTexture, PixelShaderDC, PixelShaderEmptyDC, PixelShaderSolid).

Refactored the centroid code to return a centroid. The light position is now
being converted to a point and then projected to screen. It still doesn't match
the centroid of the star; but it's pretty close and it doesn't have any magic
numbers (like 2.35) in it. I'm going to use that.