Suns are now automatically associated with XWA lights.
However, most XWA lights are plain white, so it doesn't look very interesting.
The coronas can now be colored, though.

Also added a global vector to keep pointers to textures for easy access. I'm
using this vector to store pointers to the Sun textures; but I'll also add
pointers to textures with materials, so that materials can be re-loaded while
flying.