Count the number of regions erased so that we don't display the floating HUD.
In VR, the text associated with the triangle pointer will appear in certain
areas because it's still in the floating HUD. That's annoying. This code will
remove the floating HUD if 8 regions were erased. Yeah, we might lose the beam
weapon if it isn't present in the cockpit; but oh well, let's just remember
to always put the beam weapon in the cockpit somewhere.