Fixed RenderBracket so that it doesn't render to the DCTextBuffer.
The bracket was being captured as part of the DC buffer when it shouldn't be.
I've created a separate d2d RTV just for that; but it will need a lot more work
to get it running properly in (Steam)VR mode -- so the associated hook is only
enabled if the game doesn't start in VR mode.