Experimental support for reloading materials while flying.
If reload_materials_enabled is set to 1 in SSAO.cfg, then we'll be able to
reload material definitions while flying by pressing Ctrl+Alt+L. There are
probably a number of things that can break when we reload the materials, so this
option is disabled by default and users will have to set this manually in SSAO.cfg
if they want to use this feature.

However, this is good enough for people who know what they're doing and they
just want to try out different material settings on-the-fly.