Text toggle and Highlight color for DC + minor name refactor.
bIsAimingHUd is now  bIsReticle. The intensity field for DC is now a 2-bit
integer to make more room in the alpha component. The alpha component for the
uv_coords color now has 1 bit for texture enable/disable. Added new background
highlight color that is used when the target can be fired upon. The new format
is:

uv_coords = SRC, x0,y0, x1,y1; bgColor; Intensity; TextEnable; hgColor

The alpha component of bgColor stores Intensity and TextEnable:

Bits[0-1]: Intensity - 1 (so a value of "Intensity = 1" is stored as 0)
Bits[2]: Text Enable

To toggle the text, I had to redirect the text commands to their own layer; but
this was already being done with Jeremy's Text Renderer, so that was a minor
change.