Dynamic Cockpit fixes. Landing and No-HUD problems.
When landing an X-Wing, if the Bloom effect and the new Text Rendered are
enabled, sometimes we won't resolve _offscreenBuffer to _backBuffer properly.
I still don't know *why*; but I modified the code that renders DrawHUDVertices
so that it doesn't run if the external camera + hangar + new text renderer
are enabled. Seems to work.

Second fix: if we hide all the HUD regions, then we won't clear the DC buffers
properly anymore. I added a second clause in PrimarySurface, in the Present()
path that wil clear the DC RTVs if they weren't cleared during the regular
Execute().