Added DEFAULT_FOCAL_DIST factor to the back-projection in the shaders.
The C code used to project/backproject uses an additional factor called
DEFAULT_FOCAL_DIST which is basically 2.0. This value provides an almost perfect
1:1 match with the original models. However, none of the back-projection
formulas in the shaders used this factor. If I want to use proper Shadow Mapping
I need to have the shaders use the same reconstruction or things won't match.

This change may have had some effect in a number of shaders and I may need to
check and tweak a bunch of things after this.