Refactored ShadowMap code into its own function. POV is now used from XWA's heap.
The POV is read from XWA's heap. Turns out the short* pointers are not the
current ship's POV. It seems to be a table of craft POVs. We need to use the
float* pointers and invert Y to get the current POV.

The ShadowMap code was refactored into its own function.