Fixes to Multiple Shadow Maps. OBJrange/minZ need to be per-light.
Also refactored some code out of the main light loop to avoid re-computing some
materials in every loop.

The shadows are computed before the main light loop -- not sure if this is the
best way to do it.

Shadows seem to be correct at this point; but there's still room for
improvement.