Real Vert/Horz FOV used to discard lights that aren't suns.
xwahacker's FOV is wrong. The in-game height and width should first be divided
by two and then divided by the focal length. This real FOV was confirmed by
looking at the angles the lights make with respect to the forward view. I'm
using that angle to discard lights that aren't associated with Suns.