Enabled Shadows in SSDO.
The RenderShadowMapOBJ function now sets the VS and PS constant buffers for all
modes. If shadow mapping is disabled, or if it isn't rendered (for instance,
because the camera is in exterior view), then Shadow Mapping is set to false and
the PS constant buffer is set -- this prevents artifacts.

For SSDO, we either need to multiply the shadows with SSDO or make SSDO even
darker. Otherwise, SSDO isn't quite visible...