Fixed Additional Geom Shader: it must use the in-game aspect ratio.
By co-incidence, the in-game resolution matched the aspect ratio from VRParams
and things were working properly. This didn't hold in other resolutions. The
fix is to use the current in-game aspect ratio instead of the one from
VRParams

The additional geom shader now seems to work properly on all resolutions and
FOVs.