Shadow Mapping enabled in DirectSBS mode.
However, the OBJ scale and range are wrong: the shadow maps are clipped. These
values need to be computed differently in SBS.

Some lights are still misclassified; but I'm not sure if this is only in VR
or if there's something else going on.

Also need to check the aspect_ratio and the different FOVs and in-game
resolutions.