When tagging XWA Lights, we need to use a narrower range.
Using 0.85 as the range for the FOV isn't good enough for lower in-game
resolutions. In these cases, the sun was mis-tagged as non-sun because the
centroid wasn't yet visible. I had to lower the range to 0.6 for these cases.

Part of the problem is that lights and sun textures aren't yet 100% aligned.